requirements.txt
django
djangorestframework
python-dotenv
djangorestframework-simplejwt
psycopg2
